skip wget if data is found

+++ download.sh
@@ -5,9 +5,13 @@ mkdir $DATA_DIR
 
 # Download SQuAD
 SQUAD_DIR=$DATA_DIR/squad
-mkdir $SQUAD_DIR
-wget https://rajpurkar.github.io/SQuAD-explorer/dataset/train-v1.1.json -O $SQUAD_DIR/train-v1.1.json
-wget https://rajpurkar.github.io/SQuAD-explorer/dataset/dev-v1.1.json -O $SQUAD_DIR/dev-v1.1.json
+mkdir -p $SQUAD_DIR
+if [ ! -f $SQUAD_DIR/train-v1.1.json ]; then
+	wget https://rajpurkar.github.io/SQuAD-explorer/dataset/train-v1.1.json -O $SQUAD_DIR/train-v1.1.json
+fi
+if [ ! -f $SQUAD_DIR/dev-v1.1.json ]; then
+	wget https://rajpurkar.github.io/SQuAD-explorer/dataset/dev-v1.1.json -O $SQUAD_DIR/dev-v1.1.json
+fi
 
 
 # Download CNN and DailyMail
@@ -16,10 +20,13 @@ wget https://rajpurkar.github.io/SQuAD-explorer/dataset/dev-v1.1.json -O $SQUAD_
 
 # Download GloVe
 GLOVE_DIR=$DATA_DIR/glove
-mkdir $GLOVE_DIR
-wget http://nlp.stanford.edu/data/glove.6B.zip -O $GLOVE_DIR/glove.6B.zip
+mkdir -p $GLOVE_DIR
+if [ ! -f $GLOVE_DIR/glove.6B.zip ]; then
+	wget http://nlp.stanford.edu/data/glove.6B.zip -O $GLOVE_DIR/glove.6B.zip
+fi
 unzip $GLOVE_DIR/glove.6B.zip -d $GLOVE_DIR
 
 # Download NLTK (for tokenizer)
 # Make sure that nltk is installed!
 python3 -m nltk.downloader -d $HOME/nltk_data punkt
+
